requirement.txt
torch
flask
sentence-transformers
faiss-cpu
numpy
pandas
transformers